#19 refactor: 구독 취소 UI 구성요소 컴포넌트로 분리
