+++ components/feeds/common/Button.tsx
@@ -9,7 +9,7 @@ export default function Button({ customStyle, children, type, clickHandler }: Pr
   return (
     <button
       type={type}
-      className={`mr-4 px-3 py-2 rounded-md shadow-md bg-neutral-100 text-xs text-neutral-700 dark:shadow-zinc-600 dark:bg-neutral-700 dark:text-neutral-200 ${customStyle}`}
+      className={`px-3 py-2 rounded-md shadow-md bg-neutral-100 text-xs text-neutral-700 dark:shadow-zinc-600 dark:bg-neutral-700 dark:text-neutral-200 ${customStyle}`}
       onClick={clickHandler}
     >
       {children}
